Add item frame filter expansion (#4796)
Filter items in glow item frames are expanded into a list of items if
they are a bundle, shulker box, or similar item with non-empty content.

- Hopperhock, Rannuncarpus, and Interceptor filter for the exact
container item *with the same contents* if it's in a regular item frame,
and for the contents items if it's in a glow item frame
- Corporea Funnel treats item counts of expanded filter lists as random
weights, similar to having multiple frames on the funnel block. Rotation
of the item still defines the number of items to request.

---------

Co-authored-by: Artemis System <theArtemisSystem@gmail.com>